Additional PR comments - restore default controller in `EmbeddedViewSample.Web` - remove fallback conventional route in `InlineConstraintSample.Web`
Also
- merge unique tests from `EmbeddedViewSampleCaseSensitivityTest` into `EmbeddedViewSampleTest`
- rename `area-exists` prefix since it had nothing to do with areas
